requirements.txt
streamlit
pandas
requests
#dotenv
python-dotenv
